Add elecom config to hyprland
This config makes it so that holding down the `fn3` button on the mouse
causes the trackball to scroll the page. I was able to find the button
id for `fn3` by running `wev` and seeing what id popped up when I
pressed that button.

When using the trackball for scrolling I noticed that I was scrolling in
wrong directions. I think the trackball feels more like a trackpad for
me and I fell back to the reverse scrolling method that I used to use on
mac devices. For now going to try out natural scrolling and see if it
holds true that I think that way long term.