Split home manager module by host
This setup is a step in the direction of supporting each host being able
to pick what features they want to pull in and use as their definition.
At first I was thinking all hosts would be the same but that will change
if I put this setup on a work machine.