Set up PAM on trinity
It turns out that the u2f keys are unique per host. This meant I had to
generate new keys for trinity. This causes a new level of customization
per host to be needed so now hosts nixos config is located under
`/hosts/$HOSTNAME` as opposed to the old `/hosts/$HOSTNAME.nix`.