Fix paste over top of selection vim mapping
The old approach would have issues when pasting over a selection that
had newlines. By doing a delete it would pull lines up and then paste
down and show up a line lower than expected. This new solution does a
change so that the current line can remain and the selection pasted into
it.