Hide .envrc.private over .envrc
As I have used direnv more and more I have realized that being able to
commit the `.envrc` has some advantages to it. Instead of putting secret
values in that file I am going to move to having it be a direnv entry
point that is safe to commit to repos and it will source the
`.envrc.private` file where secrets can be kept if needed.