Add TaskFS
This uses a script and envrc to make it easier to jump around and work
in taskwarrior projects. It includes the task-fs command which creates a
folder hierarchy to match the current taskwarrior projects as well as
dropping envrc files at each level to set the project. As well as some
new taskwarrior aliases to make working in the tree easier